+++ src/components/organisms/CardCompany/CardCompany.tsx
