+++ src/pages/Payments/stepper/AddCard.tsx
@@ -7,11 +7,11 @@ import Button from '@components/atoms/Button/Button';
 import Box from '@components/atoms/Box/Box';
 import Header from '@components/organisms/Header/Header';
 import CardDisplay from '@components/organisms/CardDisplay/CardDisplay';
-import CardHolderName from '@components/organisms/CardForm/CardHolderName';
-import CardNumber from '@components/organisms/CardForm/CardNumber';
-import ExpirationDate from '@components/organisms/CardForm/ExpirationDate';
-import PinNumber from '@components/organisms/CardForm/PinNumber';
-import VerificationCode from '@components/organisms/CardForm/VerificationCode';
+import CardHolderName from '@components/organisms/CardHolderName/CardHolderName';
+import CardNumber from '@components/organisms/CardNumber/CardNumber';
+import ExpirationDate from '@components/organisms/ExpirationDate/ExpirationDate';
+import PinNumber from '@components/organisms/PinNumber/PinNumber';
+import VerificationCode from '@components/organisms/VerificationCode/VerificationCode';
 import CardCompanySelect from '@components/organisms/CardCompanySelect/CardCompanySelect';
 import { cardValidate } from '@/utils/cardValidations';
 import { FormType } from '@/type/formType';
